Initial commit of PCF
This is not yet working due to not being able to define the least upper
bound of two Interp arrows. For this reason the tests of the sign and
interval semantics have also not been run on this new version where the
shared interpreter abstracts over value types.

As you can see in pcf/src/Shared.hs, we can also remove the definition
of ArrowFail with the one from lib/src/Data/Fail.hs, perhaps extract
ArrowFix into its own file in lib/src/Data as well and perhaps remove a
functional dependency in IsEnv.