GrammarSemantics: implement ArrowFix'
In order to implement ArrowFix', we need a special `top` grammar. In
the most precise case, this is the grammar over the signature of the
current subject grammar that can produce all possible terms. There
exists the `wildcard` function in the tree automata library for this
purpose, but this function (naturally) requires a signature over which
to define the grammar. In order to retrieve this signature, we need
access to the current subject grammar, which we do not have.

Hence, we create a wrapper type around `Grammar` (conveniently named
`Term`), which has a special `Top` constructor to represent this top
element. We lose precision and further complicate the code, but it works.