GrammarSemantics: use polymorphic GrammarBuilder
This commit makes the grammar semantics work with the polymorphic
`GrammarBuilder a` added in tree-automata commit 467f486. This allows
more expressitivity and therefore increases precision.

In addition, it

* increases precision of `matchTermAgainstConstructor`, by basically
taking the intersection of the grammar describing the input tuple and
the subject grammar.

* fixes the implementation for `equal`, by taking the intersection of
the two grammars. The return grammar thus generates all terms that
could _potentially_ be equal, instead of testing for grammar equality
as done originally. It also attemps to specialize terms as much as
possible now.

* fixes `mapSubterms`, in that it now only iterates those grammars
that are _immediate_ subterms of the top-level term. It now also
properly reconstructs the top-level term. For example:

  `mapSubterms f (foo[c,d])` now calls `f` on `[c,d]` and reconstructs
  a `foo(c',d')` term from the resulting terms `c'` and `d'`.

* fixes matching number- and string literals, by checking if the grammar
is a singleton grammar or not and whether it can produce the given
number- or string literal.

TODO:

1. How to treat Stratego builtins? For example, the PCF signature
references the `String` constructor but does not define it:

```
signature
  constructors
    Var  : String -> Exp
    App  : Exp * Exp -> Exp
    Abs  : String * Type * Exp -> Exp
    Zero : Exp
    Succ : Exp -> Exp
    Pred : Exp -> Exp
    Ifz  : Exp * Exp * Exp -> Exp

    Num  : Type
    Fun  : Type * Type -> Type

    Nil  : List(a)
    Cons : a * List(a) -> List(a)
         : ()
         : a -> (a)
         : a * b -> (a, b)
         : a * b * c -> (a, b, c)
```

2. Is the `Interp () a` instance for LowerBounded correct?
3. Check if production to StringLit/NumLit have empty argument list?
This should be the case by design.