GrammarSemantics: use Completion arrow transformer
This alleviates the need to define a `Complete [Term]` instance, which
does not make sense. The requirement for this instance comes from
`eval'`, which requires `Arrowtry (t,[t]) (t,[t]) (t,[t]) c` due to
`some` and `all`.

We circumvent this now by using a new `Completion` arrow transformer,
whose `Complete` instance requires a `Complete (FreeComplete y)`
instance such that there is no requirement for a `Complete y` instance.

With this, we can derive a `Complete` instance for the `Interp a b`
arrow that does not require `Complete b`, but rather `PreOrd b`.