Fixpoint: fix performance problem in grammar semantics
The problem was that `lubA (cons <<< second (arr snd) <<<
checkConstructorAndLength "Zero" []) -< toSubterms t'` does many joins.
Each `⊔` joins the underlying fixpoint cache and with this the contained
grammar. These many joins on the same grammar caused the problems. I
changed the implementation for `⊔` for `Fixpoint`, such that it does not
join the fixpoint caches anymore, but instead threads them through.

 ```
Fixpoint f ⊔ Fixpoint g = Fixpoint $ \w -> proc ((i,o),x) -> do
  (o',t1) <- f w -< ((i,o),x)
  (o'',t2) <- g w -< ((i,o'),x)
  returnA -< (o'',case (t1,t2) of
    (Terminating y',Terminating v') -> Terminating (y' ⊔ v')
    (Terminating y',NonTerminating) -> Terminating y'
    (NonTerminating,Terminating v') -> Terminating v'
    (NonTerminating,NonTerminating) -> NonTerminating)
```

This is *not* the least upper bound, because the cache is first passed
to `f` and then to `g`, but is ok because `f` an `g` increase the
fixpoint cache monotonically.